policy_server/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "policy_server"
version = "0.1.0"
description = "GPU-side policy inference server for SO-101 (LeRobot). Supports ZMQ and gRPC transports."
readme = {text = "GPU-side policy inference server that loads LeRobot policies and serves action predictions over ZMQ or gRPC.", content-type = "text/plain"}
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Dmitri Manajev", email = "dmitri@manajev.com"},
]

dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "pyzmq>=25.0",
    "msgpack>=1.0",
    "grpcio>=1.60",
]

[project.optional-dependencies]
# lerobot must be installed separately (from source or git):
#   pip install -e ./lerobot
#   pip install "lerobot[async]"
#   pip install "lerobot[smolvla]"

[project.scripts]
policy-server = "policy_server.server:main"

[tool.setuptools]
# The policy_server package lives in "." relative to this pyproject.toml
package-dir = {"policy_server" = "."}
packages = ["policy_server"]

